+++ R/bwtools.R
@@ -19,23 +19,34 @@
 #' @param bwfile bigWig file
 #' @param default_na Value to replace missing values.
 #' @importFrom dplyr mutate select `%>%`
-#' @importFrom rtracklayer BigWigFile
 #' @export
 bw_global_coverage <- function(bwfile, default_na = NA_real_) {
+    df <- bw_chr_coverage(bwfile, default_na = default_na)
+    if (!is.null(df)) {
+        result <- sum(df %>%
+                          mutate(weighted=.data[["score"]]*.data[["width"]]) %>%
+                          select("weighted")
+        ) / sum(df$width)
+        result
+    }
+}
+
+#' Calculate bigWig coverage per chromosome
+#' @param bwfile bigWig file
+#' @param default_na Value to replace missing values.
+#' @importFrom dplyr mutate select `%>%`
+#' @importFrom rtracklayer BigWigFile
+#' @export
+bw_chr_coverage <- function(bwfile, default_na = NA_real_) {
     bw <- .fetch_bigwig(bwfile)
     if (!is.null(bw)) {
         explicit_summary <- getMethod("summary", "BigWigFile")
         df <- data.frame(
             unlist(explicit_summary(bw, type="mean", default_na = default_na))
         )
-        result <- sum(df %>%
-                          mutate(weighted=.data[["score"]]*.data[["width"]]) %>%
-                          select("weighted")
-                      ) / sum(df$width)
-        result
+        df
     }
 }
-
 
 #' Score a bigWig file list and a BED file or GRanges object.
 #'
@@ -596,6 +607,50 @@ keep_canonical <- function(gr) {
     # On GRanges objects this will work as coarse, where:
     # > Remove the elements in x where the seqlevels to drop are in use.
     GenomeInfoDb::keepSeqlevels(gr, slevels, pruning.mode = "tidy")
+}
+
+#' Estimate the number of reads in a locus
+#'
+#' This is a helper tool to back-calculate a number that represents the raw
+#' counts in a locus. It is an estimate so one should consider several factors:
+#' a) fraglen parameter needs to be accurate; b) fraglen is a constant, so if
+#' the original distribution of fragment length was very heterogeneous, it
+#' will over/under estimate locus where fragments are shorter or longer than
+#' the average. c) If there was a scaling done in the original bigWig, this
+#' number will not have a 1 to 1 correspondence with the original number of
+#' reads.
+#'
+#' @param mean_cov Mean coverage
+#' @param width Width of the locus
+#' @param fraglen Fragment length
+#'
+#' @return An integer representing the estimated number of reads
+#'   (mean_cov*width) / fragment_length
+#' @export
+#'
+estimate_read_counts <- function(mean_cov, width, fraglen) {
+    round((mean_cov*width)/fraglen)
+}
+
+
+#' Convert a GRanges object with coverage values to estimated counts
+#'
+#' This method uses estimate_read_counts across a GRanges object and converts
+#' the values to read counts.
+#'
+#' @param gr GRanges object
+#' @param mcol_names Columns to calculate on
+#' @param fraglen Fragment length to use for the estimate
+#'
+#' @return A GRanges object with count values.
+#' @importFrom dplyr all_of mutate across
+#' @export
+#'
+#' @examples
+gr_coverage_to_read_counts <- function(gr, mcol_names, fraglen) {
+  df <- data.frame(gr) |>
+    mutate(across(all_of(mcol_names), ~ estimate_read_counts(.x, width, fraglen)))
+  GenomicRanges::makeGRangesFromDataFrame(df, keep.extra.columns = TRUE)
 }
 
 # Helpers ---------------------------------------------------