+++ NEWS.md
@@ -9,6 +9,17 @@ leading to inconsistent behavior. Now all the main functions with labels
 check for uniqueness in a more consistent way and throw an informative 
 warning when this requirement is not met.
 
+# wigglescout 0.21.0
+
+Date 03/06/2026
+
+## Features
+
+* `gr_coverage_to_counts` function to convert a `GRanges` object calculated by
+mean coverage into an estimate of raw counts.
+* `bw_chr_coverage` function to calculate a global mean per chromosome or
+reference present in a bigWig file. 
+
 # wigglescout 0.20.0
 
 Date: 09/03/2026